Refactoring:: 폴더로 구조화
