+++ src/pages/history.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import HistoryBox from "../components/HistoryBox";
+import HistoryBox from "../components/Profile/HistoryBox";
 import ProfileView from "@/components/Profile/ProfileView";
 import HistoryPage from "@/pageContainer/HistoryPage";
